+++ pkgs/test/lib/src/bootstrap/browser.dart
@@ -11,11 +11,10 @@ import '../runner/browser/post_message_channel.dart';
 /// Bootstraps a browser test to communicate with the test runner.
 void internalBootstrapBrowserTest(Function Function() getMain,
     {StreamChannel<Object?>? testChannel}) {
-  var channel =
-      serializeSuite(getMain, hidePrints: false, beforeLoad: () async {
-    var serialized =
-        await suiteChannel('test.browser.mapper').stream.first as Map?;
-    if (serialized == null) return;
+  var channel = serializeSuite(getMain, hidePrints: false,
+      beforeLoad: (suiteChannel) async {
+    var serialized = await suiteChannel('test.browser.mapper').stream.first;
+    if (serialized is! Map) return;
     setStackTraceMapper(JSStackTraceMapper.deserialize(serialized)!);
   });
   (testChannel ?? postMessageChannel()).pipe(channel);