+++ pkgs/test_core/lib/src/runner/vm/platform.dart
@@ -7,6 +7,7 @@ import 'dart:developer';
 import 'dart:io';
 import 'dart:isolate';
 
+import 'package:async/async.dart';
 import 'package:coverage/coverage.dart';
 import 'package:path/path.dart' as p;
 import 'package:stream_channel/isolate_channel.dart';
@@ -14,6 +15,7 @@ import 'package:stream_channel/stream_channel.dart';
 import 'package:test_api/backend.dart'; // ignore: deprecated_member_use
 import 'package:test_api/src/backend/runtime.dart'; // ignore: implementation_imports
 import 'package:test_api/src/backend/suite_platform.dart'; // ignore: implementation_imports
+import 'package:test_core/src/runner/vm/test_compiler.dart';
 import 'package:vm_service/vm_service.dart' hide Isolate;
 import 'package:vm_service/vm_service_io.dart';
 
@@ -25,6 +27,7 @@ import '../../runner/plugin/platform_helpers.dart';
 import '../../runner/runner_suite.dart';
 import '../../runner/suite.dart';
 import '../../util/dart.dart' as dart;
+import '../../util/package_config.dart';
 import '../package_version.dart';
 import 'environment.dart';
 
@@ -32,6 +35,9 @@ import 'environment.dart';
 class VMPlatform extends PlatformPlugin {
   /// The test runner configuration.
   final _config = Configuration.current;
+  final _compiler =
+      TestCompiler(p.join(p.current, '.dart_tool', 'pkg_test_kernel.bin'));
+  final _closeMemo = AsyncMemoizer<void>();
 
   VMPlatform();
 
@@ -40,15 +46,16 @@ class VMPlatform extends PlatformPlugin {
       throw UnimplementedError();
 
   @override
-  Future<RunnerSuite> load(String path, SuitePlatform platform,
-      SuiteConfiguration suiteConfig, Object message) async {
+  Future<RunnerSuite?> load(String path, SuitePlatform platform,
+      SuiteConfiguration suiteConfig, Map<String, Object?> message) async {
     assert(platform.runtime == Runtime.vm);
 
     var receivePort = ReceivePort();
-    Isolate isolate;
+    Isolate? isolate;
     try {
       isolate =
           await _spawnIsolate(path, receivePort.sendPort, suiteConfig.metadata);
+      if (isolate == null) return null;
     } catch (error) {
       receivePort.close();
       rethrow;
@@ -58,7 +65,7 @@ class VMPlatform extends PlatformPlugin {
     StreamSubscription<Event>? eventSub;
     var channel = IsolateChannel.connectReceive(receivePort)
         .transformStream(StreamTransformer.fromHandlers(handleDone: (sink) {
-      isolate.kill();
+      isolate!.kill();
       eventSub?.cancel();
       client?.dispose();
       sink.close();
@@ -108,20 +115,44 @@ class VMPlatform extends PlatformPlugin {
     return await controller.suite;
   }
 
+  @override
+  Future close() => _closeMemo.runOnce(() => _compiler.dispose());
+
   /// Spawns an isolate and passes it [message].
   ///
   /// This isolate connects an [IsolateChannel] to [message] and sends the
   /// serialized tests over that channel.
-  Future<Isolate> _spawnIsolate(
+  Future<Isolate?> _spawnIsolate(
       String path, SendPort message, Metadata suiteMetadata) async {
-    var precompiledPath = _config.suiteDefaults.precompiledPath;
-    if (precompiledPath != null) {
-      return _spawnPrecompiledIsolate(path, message, precompiledPath);
-    } else if (_config.pubServeUrl != null) {
-      return _spawnPubServeIsolate(path, message, _config.pubServeUrl!);
-    } else {
-      return _spawnDataIsolate(path, message, suiteMetadata);
+    try {
+      var precompiledPath = _config.suiteDefaults.precompiledPath;
+      if (precompiledPath != null) {
+        return _spawnPrecompiledIsolate(path, message, precompiledPath);
+      } else if (_config.pubServeUrl != null) {
+        return _spawnPubServeIsolate(path, message, _config.pubServeUrl!);
+      } else if (_config.useDataIsolateStrategy) {
+        return _spawnDataIsolate(path, message, suiteMetadata);
+      } else {
+        return _spawnKernelIsolate(path, message, suiteMetadata);
+      }
+    } catch (_) {
+      if (_closeMemo.hasRun) return null;
+      rethrow;
     }
+  }
+
+  /// Compiles [path] to kernel using [_compiler] and spawns that in an
+  /// isolate.
+  Future<Isolate> _spawnKernelIsolate(
+      String path, SendPort message, Metadata suiteMetadata) async {
+    final response =
+        await _compiler.compile(File(path).absolute.uri, suiteMetadata);
+    var compiledDill = response.kernelOutputUri?.toFilePath();
+    if (compiledDill == null || response.errorCount > 0) {
+      throw LoadException(path, response.compilerOutput ?? 'unknown error');
+    }
+    return await Isolate.spawnUri(p.toUri(compiledDill), [], message,
+        packageConfig: await packageConfigUri, checked: true);
   }
 }
 
@@ -130,11 +161,8 @@ Future<Isolate> _spawnDataIsolate(
   return await dart.runInIsolate('''
     ${suiteMetadata.languageVersionComment ?? await rootPackageLanguageVersionComment}
     import "dart:isolate";
-
     import "package:test_core/src/bootstrap/vm.dart";
-
     import "${p.toUri(p.absolute(path))}" as test;
-
     void main(_, SendPort sendPort) {
       internalBootstrapVmTest(() => test.main, sendPort);
     }